work on jupyter-free installation
